+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 - Use `std::panic` to propagate panics from the thread pool into the async
   context, rather than triggering the Rayon panic handler.
-- Add `AsyncHandle` type that implements `Future`, which makes the
+- Add `AsyncRayonHandle` type that implements `Future`, which makes the
   `async-trait` crate unnecessary.
 - Bypass Tokio `RecvError`. We control the `Sender`, so it should never be
   dropped too early.